Chrome: Refactor PostSchedule component and split it into reusable components
